aaa-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
httpx==0.27.2
python-jose==3.3.0
cachetools==5.5.0
tenacity==9.0.0
fusionauth-client==1.50.0